refactor!: move get_shared_loop in aio module
- add run_coro function
