+++ src/redsun/engine/_wrapper.py
@@ -1,20 +1,16 @@
-"""Wrapper for the [`bluesky.run_engine.RunEngine`]() class to allow execution without blocking the main thread."""
-
 from __future__ import annotations
 
 import asyncio
 from concurrent.futures import Future, ThreadPoolExecutor
 from functools import partial
-from threading import Thread
 from typing import TYPE_CHECKING, Any
 
 from bluesky.run_engine import (
     RunEngine as BlueskyRunEngine,
 )
-from bluesky.run_engine import (
-    RunEngineResult,
-    _ensure_event_loop_running,
-)
+from bluesky.run_engine import RunEngineResult
+
+from redsun.utils.aio import get_shared_loop
 
 if TYPE_CHECKING:
     from collections.abc import Awaitable, Callable, Mapping
@@ -35,37 +31,6 @@ if TYPE_CHECKING:
 def default_scan_id_source(md: dict[str, Any]) -> int:
     scan_id: int = md.get("scan_id", 0)
     return scan_id + 1
-
-
-_shared_loop: asyncio.AbstractEventLoop | None = None
-_loop_thread: Thread | None = None
-
-
-def get_shared_loop() -> asyncio.AbstractEventLoop:
-    """Return the background event loop.
-
-    Returns
-    -------
-    asyncio.AbstractEventLoop
-        The shared event loop.
-    """
-    global _shared_loop
-    global _loop_thread
-    if _shared_loop is None:
-        # at first call of this function,
-        # creates a new event loop and starts it in a background thread;
-        # subsequent calls will return the same event loop.
-        # this will be the same event loop of all
-        # RunEngine instances that use the default value of the loop parameter.
-        _shared_loop = asyncio.new_event_loop()
-        _loop_thread = Thread(target=_shared_loop.run_forever, daemon=True)
-        _loop_thread.start()
-
-        # this is a hack to make sure that the internal function
-        # that caches the event loop associated with the current thread
-        # is already aware of the loop we just created
-        _ensure_event_loop_running.loop_to_thread[_shared_loop] = _loop_thread  # type: ignore
-    return _shared_loop
 
 
 __all__ = ["RunEngine", "RunEngineResult", "get_shared_loop", "register_bound_command"]
@@ -206,6 +171,11 @@ class RunEngine(BlueskyRunEngine):
 
     """
 
+    # TODO: using get_shared_loop() like this is a bit
+    # fragile; there should be a private function that ensures
+    # the shared loop is created only once at application startup
+    # and properly cleaned up at shutdown; this is just a quick solution to
+    # get the shared loop working for now
     def __init__(
         self,
         md: dict[str, Any] | None = None,
